Previne IsNaN e Salva carrinho em Storage, inclusive atualiza o carrinho em mais de uma aba aberta

+++ script.js
@@ -69,8 +69,11 @@ const getSectionTag = (group) => {
 }
 
 // ADD PRODUCTS TO CART
-
 let addProductsCart = []
+const productInStorege = localStorage.getItem('productsinCart')
+if (productInStorege) {
+  addProductsCart = JSON.parse(productInStorege)
+}
 const addToCart = newProduct => {
   const productIndex = addProductsCart.findIndex(
     item => item.id === newProduct.id
@@ -83,19 +86,31 @@ const addToCart = newProduct => {
   } else {
     addProductsCart[productIndex].qty++
   }
-  CartUpdate()
+  CartUpdate(true)
 }
 
 const inputQtyUpdate = (id, newQty) => {
+  const newQtyInt = parseInt(newQty)
+  if (isNaN(newQtyInt)){
+    return
+  }
+  // if (newQtyInt > 10){
+  //   alert('Não é permitido iserir um número maior que 10')
+  //   return
+  // }
+  if (newQtyInt > 0){
   const findIndex = addProductsCart.findIndex((product) => {
     if (product.id === id){
       return true
     }
       return false
-    
+      
     })
-  addProductsCart[findIndex].qty = parseInt(newQty)
-  CartUpdate()
+    addProductsCart[findIndex].qty = parseInt(newQty)
+    CartUpdate(false)
+}else{
+  deleteToCart(id)
+} 
 }
 
 const deleteToCart = (id) => {
@@ -108,13 +123,17 @@ const deleteToCart = (id) => {
   addProductsCart = addProductsCart.filter(    
     product => product.id != id
     )
-  CartUpdate()
+  CartUpdate(true)
   if(addProductsCart.length === 0){
     closeCart()
   }
   }
 
-const CartUpdate = () => {
+const CartUpdate = (printProducts) => {
+  //SAVE CART IN A LOCALSTORAGE
+  const productsCartSotorege = JSON.stringify(addProductsCart)
+  localStorage.setItem('productsinCart', productsCartSotorege)
+  //VARIABLES
   const selectDivEmptyCart = document.querySelector ('.empty-cart')
   const cartBadge = document.querySelector('.btn-cart-badge')
   const selectBtnRequest = document.querySelector ('.btn-request')
@@ -129,39 +148,40 @@ const CartUpdate = () => {
     })
     cartBadge.textContent = total
     //UPDATE CART
-    selectAside.classList.add('add-products-cart-on')
-    selectBtnRequest.classList.add('btn-request-on')
-    selectAsideUl.innerHTML=''
-    addProductsCart.forEach(product => {
-      const createLi = document.createElement('li')
-        createLi.classList.add('product-select')
-        createLi.innerHTML = `
-        <img class="image-cart" src="${product.image}" alt="${product.name}" height="40" width="40" />
-        <p>${product.name}</p>
-        <input class="inputs" type="number" value="${product.qty}"/>
-        <button><img src="images/trash.svg" alt="Excluir" height="27" width="22" /></button>
-      `
-      const selectInput = createLi.querySelector('input')
-      selectInput.addEventListener('keyup', (event) => {
-        inputQtyUpdate(product.id, event.target.value)
-      })
-      selectInput.addEventListener('change', (event) => {
-        inputQtyUpdate(product.id, event.target.value)
-      })
-      selectInput.addEventListener('keydown', (event) => {
-        if (event.key === '.' || event.key === ',' || event.key === '-'){
-          event.preventDefault()
-        }
-      })
-
-      const selectBtnDelete = createLi.querySelector('button')
-      selectBtnDelete.addEventListener('click', () => {
-        deleteToCart(product.id)
-      })
-      selectAsideUl.appendChild(createLi)
-  })
-    //TURN-OFF EMPTY CART
-    selectDivEmptyCart.classList.remove('empty-cart-on')
+    if (printProducts){
+      selectAside.classList.add('add-products-cart-on')
+      selectBtnRequest.classList.add('btn-request-on')
+      selectAsideUl.innerHTML=''
+      addProductsCart.forEach(product => {
+        const createLi = document.createElement('li')
+          createLi.classList.add('product-select')
+          createLi.innerHTML = `
+          <img class="image-cart" src="${product.image}" alt="${product.name}" height="40" width="40" />
+          <p>${product.name}</p>
+          <input class="inputs" type="number" value="${product.qty}"/>
+          <button><img src="images/trash.svg" alt="Excluir" height="27" width="22" /></button>
+        `
+        const selectInput = createLi.querySelector('input')
+        selectInput.addEventListener('keyup', (event) => {
+          inputQtyUpdate(product.id, event.target.value)
+        })
+        selectInput.addEventListener('change', (event) => {
+          inputQtyUpdate(product.id, event.target.value)
+        })
+        selectInput.addEventListener('keydown', (event) => {
+          if (event.key === '.' || event.key === ',' || event.key === '-'){
+            event.preventDefault()
+          }
+        }) 
+        const selectBtnDelete = createLi.querySelector('button')
+        selectBtnDelete.addEventListener('click', () => {
+          deleteToCart(product.id)
+        })
+        selectAsideUl.appendChild(createLi)
+        })
+        //TURN-OFF EMPTY CART
+        selectDivEmptyCart.classList.remove('empty-cart-on')
+    }
   } else {
     //TURN-ON EMPTY CART
     selectDivEmptyCart.classList.add('empty-cart-on')
@@ -173,6 +193,12 @@ const CartUpdate = () => {
 
   }
 }
+CartUpdate(true)
 
-CartUpdate()
+window.addEventListener('storage', (e)=> {
+  if (e.key === 'productsinCart'){
+    addProductsCart = JSON.parse(e.newValue)
+    CartUpdate(true)
+  }
+})
